Removed Badge component usage.
