+++ packages/components/src/InlineEditInput/InlineEditInfinity/index.tsx
@@ -1,27 +1,10 @@
 import React, { useCallback } from 'react';
-import classNames from 'classnames';
 
 import { isInfinite, parseInfinity } from '@eventespresso/utils';
 import { InlineEdit } from '@eventespresso/adapters';
-import type { PreviewProps } from './types';
 
-import { TabbableText } from '../index';
-import type { TextProps } from './types';
-
-const Preview: React.FC<PreviewProps> = ({ value, onRequestEdit, isEditing, ...props }) => {
-	const isInfinity = isInfinite(value);
-	const className = classNames('ee-inline-edit__infinity', {
-		'ee-infinity-sign': isInfinity,
-	});
-
-	if (isEditing) {
-		return null;
-	}
-
-	const output = isInfinity ? <span className={'ee-infinity-sign__inner'}>{'∞'}</span> : value;
-
-	return <TabbableText {...props} bottomBorderDashed className={className} onClick={onRequestEdit} text={output} />;
-};
+import Preview from './Preview';
+import type { TextProps } from '../types';
 
 const InlineEditInfinity: React.FC<TextProps> = ({ onChangeValue, value, ...props }) => {
 	const isInfinity = isInfinite(value);
