+++ packages/components/src/InlineEditInput/style.scss
@@ -43,13 +43,20 @@
 	display: inline-flex;
 	position: relative;
 
-	.badge {
+	svg {
 		background-color: transparent;
 		height: 1.25rem;
-		width: 1.25rem;
+		margin: 0;
 		position: absolute;
 		top: 0;
-		right: -1rem;
+		right: -2rem;
+		width: 1.25rem;
 		z-index: 99;
+	}
+
+	&--infinity {
+		svg {
+			right: -22px;
+		}
 	}
 }
