+++ packages/components/src/TabbableText/types.ts
@@ -1,7 +1,6 @@
 export interface TabbableTextProps {
 	className?: string;
 	icon?: React.ReactNode;
-	bottomBorderDashed?: boolean;
 	onClick: VoidFunction;
 	richTextContent?: boolean;
 	text?: string | JSX.Element;
